refactor: remove circular dependency and add CircularDependencyRspackPlugin
