+++ packages/core/src/utils/color.ts
@@ -1,0 +1,2 @@
+import color from 'picocolors';
+export { color };
